==> CAP_AI_Repository/01_Preprocessing/00_Data_2016_2018/05_00_Mixing_eobs_haemt_2016_2018.ipynb <==
%reset -f
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import re, sqlite3, pickle, time, datetime, random, os
from dateutil.relativedelta import relativedelta

pd.options.display.float_format = '{:,.4f}'.format

# Load Processed parameters and data

##############################################################################
#######################################
### Relevant fields for pre-processing
#######################################
reference_date = datetime.datetime(1970,1,1)

##### Fields in table patients ##########
patient_field         = 'patient_id'
age_field             = 'age_at_admission\r'
sex_field             = 'sex'
ethnic_field          = 'ethnic_origin'
death_ind_field       = 'death_indicator'
death_date_field      = 'date_of_death'
death_date_code_field = 'date_code_of_death'
mort_in_hosp_field    = 'Thirty_day_mort'

##### Fields in table admissions ########
admission_field           = 'admission_id'
diagnosis_field           = 'episode_diagnoses'
admn_date_field           = 'admission_date_time'
admn_discharge_field      = 'discharge_date_time'
admn_date_code_field      = 'admission_date_code_time'
admn_discharge_code_field = 'discharge_date_code_time'
lengthofstay_field        = 'lengthofstay'
isPneumonia_field         = 'isPneumonia'
mortal_admin_field        = 'mortal_admin'
comorbidity_field         = 'Comorbidity_score'
icu_admin_field           = 'icu_count\r'
no_eobs_field             = 'no_obs_eobs'
no_haematology_field      = 'no_haematology_eobs'
meds_drug_field           = 'Drug_supplied'
specific_comorb_field     = 'Specific Comorbidity'
prev_admin_field          = 'had_Prev_admin'
has_spin_field            = 'adm_has_spin'
cubr65_field              = 'CURB65'
antibiot_4h_field         = 'antibiotic_4h'

 ##### Fields in table eObservations #####
eObs_time_field      = 'timestamp'
eObs_time_code_field = 'timestamp_code'
eObs_time_prev_obs   = 'time_since_prev_obs_in_mins'
eObs_resprate_field  = 'rr'
eObs_sbp_field       = 'sbp'
eObs_dbp_field       = 'dbp'
eObs_newscore_field  = 'ews'
eObs_heartrate_field = 'heart_rate'
eObs_temptr_field    = 'temperature\r'
eObs_oxygen_field    = 'Oxygen_Saturation'

##### Fields in table haematological tests #####
test_code_field      = 'local_test_code'
test_time_field      = 'sample_collection_date_time'
test_time_code_field = 'sample_collection_date_code_time'

data = pickle.load( open('DataFrame_pickles/df_patients_admissions_2016_18_v4.pickle', 'rb'))
df_patients   = data[0]
df_admissions = data[1]

df_eobs = pickle.load( open("DataFrame_pickles/df_eobs_3d.pickle", "rb" ))
df_haem = pickle.load( open('DataFrame_pickles/new_haematology_3d.pickle', 'rb')) 
df_haem = df_haem.rename(columns = {'sample_collection_date_time':'timestamp'})

# Selection of Blood Tests


df = pd.DataFrame(df_haem.count())
df = df.rename(columns = {0: 'Count'}).sort_values(by = ['Count'], ascending = False)
df['Per'] = df['Count'] / len(df_haem)
display(df)
BT_include = df[df['Per'] >= 0.7].index.tolist()[2:]
BT_include = ['CREA','UREA','K','GFR','WBC','PLT','HCT','HGB','RBC','MCH','MCV','NEUAB','TLYMAB',
              'EOSAB','MONAB','BASAB','ALB','ALP','BILI']

print("------------------------------------------------------")
print("Number of Blood Test to include in the merging dataset", len(BT_include))
print("")
print(BT_include)

df_haem_merge = df_haem[[admission_field, eObs_time_field] + BT_include]

display(df_haem_merge.describe())
display(df_eobs.describe())

A = df_admissions[admission_field].unique().tolist()
B = df_eobs[admission_field].unique().tolist()
C = df_haem_merge[admission_field].unique().tolist()
print("No. admissions in admissions", len(A))
print("No. admissions in eobs", len(B))
print("No. admissions in haematology", len(C))

D = set(A)&set(B)&set(C)

print("No. admissions in all", len(D))
admissions_to_process = list(D)

df = pd.merge(df_eobs[df_eobs[admission_field].isin(admissions_to_process)], df_haem_merge, on = [admission_field, eObs_time_field],how="left")

df.describe()

### Merging check Point before interpolation

admin = 67762672
admin = 52573088

print("__________________________________________________________________________________________________________")
print("ADMISSION INFORMATION")
display(df_admissions[df_admissions[admission_field]==admin])

print("")
print("")
print("=========================== time series EOBS table ====================================")
display(df_eobs[df_eobs[admission_field] == admin].iloc[:4])
print("")
print("")
print("=========================== time series HAEMATOLOGY info table ====================================")
display(df_haem_merge[df_haem_merge[admission_field] == admin].iloc[:4])
print("")
print("")
print("=========================== time series HMIXED table ====================================")
display(df[df[admission_field] == admin].iloc[:4])

print("__________________________________________________________________________________________________________")
print("")
print("df_eobs, number of records:", len(df_eobs[df_eobs[admission_field] == admin]))

print("df_haem, number of records:", len(df_haem_merge[df_haem_merge[admission_field] == admin]))

print("df_mixed number of records", len(df[df[admission_field] == admin]))


## Interpolation of Haematology Features

def interpolate_by_method(series, method, plot = 0):
    if ('polynomial' in method) or ('nearest' in method):
        interpolate = series.interpolate(method = method[:-1], order = int(method[-1]))
    else:
        interpolate = series.interpolate(method = method, order = method)
    if plot != 0: interpolate.plot()
    return interpolate

def fill_top_bottom_values(series):
    interp =series.copy()
    first_non_na = 0
    last_non_na  = 0
    for idx,val in interp.items():
        if (val != val) and (first_non_na == 0): # if val is nan and first_non_na = 0
            continue
        elif not (val != val) and (first_non_na == 0):
            first_non_na = idx       
        elif (val != val) and (first_non_na != 0) and (last_non_na == 0):
            last_non_na = idx - relativedelta(minutes = 15)
    
    if first_non_na !=0:
        for i in range(len(interp.loc[:first_non_na]) - 1):
            time_to_change  = first_non_na - relativedelta(minutes = 15 * (i+1))
            fifth_next_time = time_to_change + relativedelta(minutes = 15*10)
            interp.at[time_to_change] = interp.loc[time_to_change:fifth_next_time].mean()

    if last_non_na !=0:
        for i in range(len(interp.loc[last_non_na:])-1):
            time_to_change  = last_non_na + relativedelta(minutes = 15 * (i+1))
            fifth_prev_time = time_to_change - relativedelta(minutes = 15*10)
            interp.at[time_to_change] = interp.loc[fifth_prev_time:time_to_change].mean()
    return interp

t =time.time()
df_eobs_haem_mx = pd.DataFrame(columns = df.columns)
list_features = list(df_haem_merge.columns[2:])

for adm in df[admission_field].unique().tolist():

    df_eob_mx_adm   = df[df[admission_field] == adm].copy()
    upsamp_times    = df_eob_mx_adm[eObs_time_field].values
    
    for idx_feat, feature in enumerate(list_features):
        
        values  = df_eob_mx_adm[feature].values
        series_ = pd.Series(values, index=upsamp_times)
        if len(series_) - series_.isna().sum() < 4: continue
        interp  = interpolate_by_method(series_, 'linear')
        if len(interp) == interp.isna().sum(): continue
            
        interp = interp.fillna(method="bfill")    
        interp = interp.fillna(method="ffill")    
    
        #interp  = fill_top_bottom_values(interp)
        
        df_eob_mx_adm[feature] = interp.values
        
    df_eobs_haem_mx = pd.concat([df_eobs_haem_mx, df_eob_mx_adm])
   
print("elpased:", time.time()-t)

# --------------------------------------------------------------------
#Reduce the time series to 144 samples in the three days. which means a reading every 30 minutes. 
# --------------------------------------------------------------------
df_eobs_haem_mx = df_eobs_haem_mx.iloc[::2]
df_eobs_haem_mx.reset_index(drop=True)

# Resumen

admin = 67762672
admin = 52573088


print("__________________________________________________________________________________________________________")
print("ADMISSION INFORMATION")
display(df_admissions[df_admissions[admission_field]==admin])

print("")
print("")
print("=========================== MIX time series BEFORE interpolation ====================================")
display(df[df[admission_field] == admin].iloc[:5])

print("")
print("")
print("=========================== MIX time series AFTER interpolation ====================================")
display(df_eobs_haem_mx[df_eobs_haem_mx[admission_field] == admin].iloc[:5])

print("__________________________________________________________________________________________________________")
print("")
print("Number of readings in timeseries BEFORE interpolation", len(df[df[admission_field] == admin]))
print("Number of readings in timeseries AFTER  interpolation", len(df_eobs_haem_mx[df_eobs_haem_mx[admission_field] == admin]))

# Add Confusion to data
The confusion field is a binary field computed from the Early warning score. 
Document NEWS2 Chart 1_The NEWS scoring system from [Link](https://www.rcplondon.ac.uk/projects/outputs/national-early-warning-score-news-2)

# -----------------------------------------------------------------
# Functions to find confusion
# -----------------------------------------------------------------
def ews_rr_score(rr_row):
    rr_score = 0
    if (rr_row >= 25) or (rr_row < 9):     rr_score = 3
    elif (21 <= rr_row) and (rr_row <= 24):rr_score = 2
    elif (9 <= rr_row) and (rr_row <= 11): rr_score = 1
    else: rr_score = 0
    return rr_score
# -----------------------------------------------------------------
# Oxygen saturation Score
def ews_ox_score(ox_row, as_ox_row):
    ox_score = 0
    if (ox_row <= 91): ox_score = 3
    elif (92 <= ox_row) and (ox_row <= 93): ox_score = 2
    elif (93 < ox_row) and (ox_row <= 95): ox_score  = 1
    else: ox_score = 0        
    if as_ox_row == 1: ox_score = ox_score + 2
    return ox_score

# -----------------------------------------------------------------
# Systolic blood presure score
def ews_sbp_score(sbp_row):
    sbp_score = 0
    if (sbp_row >= 220) or (sbp_row <= 90):    sbp_score = 3
    elif (91 <= sbp_row) and (sbp_row <= 100): sbp_score = 2
    elif (101 <= sbp_row) and (sbp_row <= 110):sbp_score = 1
    else: sbp_score = 0
    return sbp_score

# -----------------------------------------------------------------
# Heart rate score
def ews_hr_score(hr_row):
    hr_score = 0
    if (hr_row >= 131) or (hr_row <= 40):    hr_score = 3
    elif (111 <= hr_row) and (hr_row <= 130): hr_score = 2
    elif (91 <= hr_row) and (hr_row <= 110):hr_score = 1
    elif (41 <= hr_row) and (hr_row <= 50): hr_score = 1
    else: hr_score = 0
    return hr_score

# -----------------------------------------------------------------
# Temperature score
def ews_temp_score(temp_row):
    temp_score = 0
    if (temp_row <= 31.5):   temp_score = 3
    elif (39.1 <= temp_row): temp_score = 2
    elif (35.1 <= temp_row) and (temp_row <= 36):temp_score = 1
    elif (38.1 <= temp_row) and (temp_row <= 39): temp_score = 1
    else: temp_score = 0
    return temp_score

# -----------------------------------------------------------------
# Compute Confusion using the information of the Row
def confusion_from_row(row):
    confusion_row = 0
    ews_row = row['ews'] 
    rr_row = row['rr']
    ox_row    = row['Oxygen_Saturation']
    as_ox_row = row['Assisted_O2']
    sbp_row    = row['sbp']
    hr_row    = row['heart_rate']
    temp_row    = row['temperature\r']
    ews_score = ews_rr_score(rr_row)
    ews_score = ews_score + ews_ox_score(ox_row, as_ox_row)
    ews_score = ews_score + ews_sbp_score(sbp_row)
    ews_score = ews_score + ews_hr_score(hr_row)
    ews_score = ews_score + ews_temp_score(temp_row)
    if (abs(ews_score - ews_row) <= 3) and (abs(ews_score - ews_row) > 0): confusion_row = 1
    return confusion_row 

t = time.time()
df_eobs_haem_mx['Confusion'] = df_eobs_haem_mx.apply(lambda x:confusion_from_row(x), axis = 1)
col = df_eobs_haem_mx.pop('Confusion')
df_eobs_haem_mx.insert(12, col.name, col)

print("time elapsed: ", time.time() - t)

df_eobs_haem_mx.columns

df_eobs_haem_mx.head(5)

# Cleanning the final mixed data frame

df_eobs_haem_mx.isna().sum()

t =time.time()
df_eobs_haem_mx_v2 = pd.DataFrame(columns = df_eobs_haem_mx.columns)
for adm in df_eobs_haem_mx[admission_field].unique().tolist():
    df_adm = df_eobs_haem_mx[df_eobs_haem_mx[admission_field] == adm].copy()
    if df_adm.isna().sum().sum() == 0:
        df_eobs_haem_mx_v2 = pd.concat([df_eobs_haem_mx_v2, df_adm])
print("elpased:", time.time()-t)   

df_eobs_haem_mx_v2.isna().sum()

df_eobs_haem_mx_v2.describe()

len(df_eobs_haem_mx_v2[admission_field].unique().tolist())

features = df_eobs_haem_mx_v2.columns[2:]
for feat in features:
    if df_eobs_haem_mx_v2[feat].dtypes == 'O':
        df_eobs_haem_mx_v2[feat] = pd.to_numeric(df_eobs_haem_mx_v2[feat])
pickle.dump([df_eobs_haem_mx_v2], open('DataFrame_pickles/df_eobs_heam_mixed_2016_2018_V2.pickle', 'wb'))

t =time.time()
df_eobs_haem_mx_v3 = pd.DataFrame(columns = df_eobs_haem_mx.columns[:15])
for adm in df_eobs_haem_mx[admission_field].unique().tolist():
    df_adm = df_eobs_haem_mx[df_eobs_haem_mx[admission_field] == adm][df_eobs_haem_mx.columns[:15]].copy()
    if df_adm.isna().sum().sum() == 0:
        df_eobs_haem_mx_v3 = pd.concat([df_eobs_haem_mx_v3, df_adm])
print("elpased:", time.time()-t)   

len(df_eobs_haem_mx_v3[admission_field].unique().tolist())

features = df_eobs_haem_mx_v3.columns[2:]
for feat in features:
    if df_eobs_haem_mx_v3[feat].dtypes == 'O':
        df_eobs_haem_mx_v3[feat] = pd.to_numeric(df_eobs_haem_mx_v3[feat])
pickle.dump([df_eobs_haem_mx_v3], open('DataFrame_pickles/df_eobs_heam_mixed_2019_2020_V3.pickle', 'wb'))
